pyproject.toml
[project]
name = "ginjer-exercice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Zawnia", email = "al84250@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.73.1",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "langfuse>=3.0.0",
    "pyyaml>=6.0",
    "typer>=0.24.1",
    "google-cloud-bigquery>=3.41.0",
    "httpx>=0.28.1",
]

[project.scripts]
ginjer-exercice = "ginjer_exercice:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "respx>=0.23.1",
    "ruff>=0.15.11",
]
